# docs/api/openapi.yaml
openapi: 3.1.2
info:
  title: Spring AWS Portfolio API
  version: 0.4.0
  summary: MVP HTTP contract for the Spring AWS Portfolio project.
  description: |
    Design-first API contract for the first-party React + Spring Boot application.

    Authentication uses a server-side Spring Session persisted through JDBC. Production
    uses an HttpOnly/Secure `__Host-SESSION` cookie. Unsafe requests require the CSRF
    token returned by `GET /api/v1/auth/csrf` in the `X-CSRF-TOKEN` header.

    Assessment scoring metadata (`dimension`, keyed pole, scoring/ambiguity rules and
    DefinitionVersion internals) is never exposed through the public questionnaire DTOs.
servers:
  - url: /
...
security:
  - SessionCookie: []
paths:
...
  /api/v1/assessment-sessions:
    get:
      tags: [Assessment Sessions]
      operationId: listAssessmentHistory
      summary: List the authenticated user's completed assessment history.
      parameters:
        - name: status
          in: query
          required: false
          description: MVP history only exposes completed assessments.
          schema:
            type: string
            enum: [COMPLETED]
            default: COMPLETED
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: Assessment history page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentHistoryPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/assessment-sessions/{sessionId}/clarifications/{dimensionCode}/skip:
    parameters:
      - $ref: '#/components/parameters/SessionIdPath'
      - $ref: '#/components/parameters/DimensionCodePath'
    post:
      tags: [Clarifications]
      operationId: skipClarification
      summary: Skip one eligible clarification and re-evaluate finalization readiness.
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200':
          description: Clarification skipped and Session state updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/assessment-sessions/{sessionId}/clarifications/skip-remaining:
    parameters:
      - $ref: '#/components/parameters/SessionIdPath'
    post:
      tags: [Clarifications]
      operationId: skipRemainingClarifications
      summary: Skip all remaining eligible clarifications.
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200':
          description: Remaining clarifications skipped and Session state updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/assessment-sessions/{sessionId}/tie-breaks/{dimensionCode}:
    parameters:
      - $ref: '#/components/parameters/SessionIdPath'
      - $ref: '#/components/parameters/DimensionCodePath'
    put:
      tags: [Assessment Sessions]
      operationId: submitDimensionTieBreak
      summary: Record the user's explicit final pole selection for a required exact tie.
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TieBreakRequest'
      responses:
        '200':
          description: Tie-break fact accepted and Session state updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-SESSION
      description: |
        Production server-side session cookie. It is HttpOnly, Secure, SameSite=Lax,
        Path=/ and has no Domain attribute. The local development profile may use the
        non-Secure name `SESSION`; the production contract is shown here.
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      description: Token obtained from `GET /api/v1/auth/csrf`; required for unsafe requests.
      schema:
        type: string
        minLength: 1
...
    SessionIdPath:
      name: sessionId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
    DimensionCodePath:
      name: dimensionCode
      in: path
      required: true
      description: Definition-driven dimension code; not restricted to the current personality assessment's four codes at transport level.
      schema:
        type: string
        minLength: 1
        maxLength: 100
    PageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeQuery:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Malformed JSON, missing/invalid transport fields, or query/path validation failure.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Authentication is missing/expired/invalid, or credentials are invalid for login.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Authenticated/anonymous request is forbidden, including CSRF validation failure where applicable.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource does not exist or is intentionally hidden from this actor.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Current business/resource state conflicts with the requested action.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    UnprocessableContent:
      description: Structurally valid input has invalid business semantics.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    InternalServerError:
      description: Unexpected server failure.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    Uuid:
      type: string
      format: uuid
    Instant:
      type: string
      format: date-time
...
    QuestionAnswer:
      type: object
      additionalProperties: false
      required: [questionId, value]
      properties:
        questionId:
          type: string
          minLength: 1
        value:
          type: integer
          description: Definition-driven canonical answer value; current personality v1 uses 1..5.
    AssessmentSessionResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - assessment
        - status
        - questionnaire
        - initialResult
        - clarifications
        - tieBreaks
        - finalResult
        - workflow
        - createdAt
        - completedAt
        - abandonedAt
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        assessment:
          $ref: '#/components/schemas/BoundAssessmentRef'
        status:
          $ref: '#/components/schemas/AssessmentSessionStatus'
        questionnaire:
          $ref: '#/components/schemas/QuestionnaireState'
        initialResult:
          anyOf:
            - $ref: '#/components/schemas/InitialAssessmentResult'
            - type: 'null'
        clarifications:
          type: array
          items:
            $ref: '#/components/schemas/ClarificationState'
        tieBreaks:
          type: array
          items:
            $ref: '#/components/schemas/TieBreakState'
        finalResult:
          anyOf:
            - $ref: '#/components/schemas/FinalAssessmentResult'
            - type: 'null'
        workflow:
          $ref: '#/components/schemas/AssessmentWorkflow'
        createdAt:
          $ref: '#/components/schemas/Instant'
        completedAt:
          anyOf:
            - $ref: '#/components/schemas/Instant'
            - type: 'null'
        abandonedAt:
          anyOf:
            - $ref: '#/components/schemas/Instant'
            - type: 'null'
    BoundAssessmentRef:
      type: object
      additionalProperties: false
      required: [code, version]
      properties:
        code:
          type: string
        version:
          type: string
    AssessmentSessionStatus:
      type: string
      enum:
        - IN_PROGRESS
        - AWAITING_CLARIFICATION
        - CLARIFICATION_IN_PROGRESS
        - COMPLETED
        - ABANDONED
    QuestionnaireState:
      type: object
      additionalProperties: false
      required: [answers, submitted, submittedAt]
      properties:
        answers:
          type: array
          items:
            $ref: '#/components/schemas/QuestionAnswer'
        submitted:
          type: boolean
        submittedAt:
          anyOf:
            - $ref: '#/components/schemas/Instant'
            - type: 'null'
    InitialAssessmentResult:
      type: object
      additionalProperties: false
      required: [dimensions]
      properties:
        dimensions:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/InitialDimensionResult'
    InitialDimensionResult:
      type: object
      additionalProperties: false
      required: [dimensionCode, rawScore, ambiguous, evidence]
      properties:
        dimensionCode:
          type: string
        rawScore:
          type: integer
        questionnairePreference:
          type: [string, 'null']
          description: Null for an exact tie.
        ambiguous:
          type: boolean
        evidence:
          $ref: '#/components/schemas/QuestionnaireEvidence'
    QuestionnaireEvidence:
      type: object
      additionalProperties: false
      required: [poleA, poleAPercentage, poleB, poleBPercentage]
      properties:
        poleA:
          type: string
        poleAPercentage:
          type: number
          minimum: 0
          maximum: 100
        poleB:
          type: string
        poleBPercentage:
          type: number
          minimum: 0
          maximum: 100
    ClarificationState:
      type: object
      additionalProperties: false
      required: [dimensionCode, status, result, startedAt, acceptedAt]
      properties:
        dimensionCode:
          type: string
        status:
          $ref: '#/components/schemas/ClarificationStatus'
        result:
          anyOf:
            - $ref: '#/components/schemas/ClarificationResult'
            - type: 'null'
        startedAt:
          anyOf:
            - $ref: '#/components/schemas/Instant'
            - type: 'null'
        acceptedAt:
          anyOf:
            - $ref: '#/components/schemas/Instant'
            - type: 'null'
    ClarificationStatus:
      type: string
      enum:
        - PENDING
        - IN_PROGRESS
        - CLARIFIED
        - SKIPPED
        - FAILED_RETRYABLE
    ClarificationResult:
      oneOf:
        - $ref: '#/components/schemas/ResolvedClarificationResult'
        - $ref: '#/components/schemas/UnclearClarificationResult'
    ResolvedClarificationResult:
      type: object
      additionalProperties: false
      required: [resolution, suggestedPole, confidence]
      properties:
        resolution:
          type: string
          const: RESOLVED
        suggestedPole:
          type: string
          minLength: 1
        confidence:
          type: string
          description: Provider/domain confidence label retained as accepted business evidence.
        reasoningSummary:
          type: [string, 'null']
          description: Optional concise product/audit summary; never model private chain-of-thought.
    UnclearClarificationResult:
      type: object
      additionalProperties: false
      required: [resolution, suggestedPole, confidence]
      properties:
        resolution:
          type: string
          const: UNCLEAR
        suggestedPole:
          type: 'null'
          description: UNCLEAR is an accepted business result and therefore has no selected pole.
        confidence:
          type: string
          description: Provider/domain confidence label retained as accepted business evidence.
        reasoningSummary:
          type: [string, 'null']
          description: Optional concise product/audit summary; never model private chain-of-thought.
    TieBreakState:
      type: object
      additionalProperties: false
      required: [dimensionCode, selectedPole, decidedAt]
      properties:
        dimensionCode:
          type: string
        selectedPole:
          type: string
        decidedAt:
          $ref: '#/components/schemas/Instant'
    FinalAssessmentResult:
      type: object
      additionalProperties: false
      required: [finalType, dimensions]
      properties:
        finalType:
          type: string
          minLength: 1
        dimensions:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/FinalDimensionConclusion'
    FinalDimensionConclusion:
      type: object
      additionalProperties: false
      required: [dimensionCode, finalPreference, source, overrodeBaseline]
      properties:
        dimensionCode:
          type: string
        questionnairePreference:
          type: [string, 'null']
        finalPreference:
          type: string
        source:
          $ref: '#/components/schemas/FinalDecisionSource'
        overrodeBaseline:
          type: boolean
    FinalDecisionSource:
      type: string
      enum:
        - QUESTIONNAIRE
        - QUESTIONNAIRE_CONFIRMED_BY_CLARIFICATION
        - AI_CLARIFICATION
        - QUESTIONNAIRE_FALLBACK
        - USER_TIE_BREAK
    AssessmentWorkflow:
      type: object
      additionalProperties: false
      required:
        - pendingClarificationDimensions
        - retryableClarificationDimensions
        - tieBreakRequiredDimensions
        - completed
      properties:
        pendingClarificationDimensions:
          type: array
          items:
            type: string
        retryableClarificationDimensions:
          type: array
          items:
            type: string
        tieBreakRequiredDimensions:
          type: array
          items:
            type: string
        completed:
          type: boolean
...
    TieBreakRequest:
      type: object
      additionalProperties: false
      required: [selectedPole]
      properties:
        selectedPole:
          type: string
          minLength: 1
          pattern: '.*\S.*'
    AssessmentHistoryItem:
      type: object
      additionalProperties: false
      required: [sessionId, assessmentCode, assessmentVersion, finalType, completedAt]
      properties:
        sessionId:
          $ref: '#/components/schemas/Uuid'
        assessmentCode:
          type: string
        assessmentVersion:
          type: string
        finalType:
          type: string
        completedAt:
          $ref: '#/components/schemas/Instant'
    AssessmentHistoryPage:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AssessmentHistoryItem'
...
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    ProblemDetails:
      type: object
      additionalProperties: true
      required: [type, title, status, code]
      properties:
        type:
          type: string
          description: Stable problem type URI/reference, e.g. `/problems/join-request-already-pending`.
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: [string, 'null']
        instance:
          type: [string, 'null']
        code:
          type: string
          pattern: '^[A-Z0-9_]+$'
          description: Stable application error code for frontend branching.
        traceId:
          type: [string, 'null']
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        retryAt:
          anyOf:
            - $ref: '#/components/schemas/Instant'
            - type: 'null'
          description: Present for errors such as JOIN_REQUEST_COOLDOWN_ACTIVE when applicable.
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string